Move wolverine-specific tests to its own class
